Split in RestUi and RestAPI
